Make it possible to use a config file instead of N arguments
