+++ config.json.sample
@@ -1,0 +1,8 @@
+{
+    "host": "127.0.1.1",
+    "port": 4141,
+    "domain": "box.knilxof.org",
+    "tunnel_ip": "1.2.3.4",
+    "dns_ttl": 89,
+    "eviction_delay": 123
+}
