+++ src/main.rs
@@ -31,14 +31,13 @@ extern crate serde_derive;
 extern crate serde_json;
 extern crate uuid;
 
-use clap::{App, ArgMatches};
 use hyper_openssl::OpensslServer;
 use iron::{Chain, Iron};
 use iron::method::Method;
 use iron_cors::CORS;
 use mount::Mount;
-use std::path::PathBuf;
 
+mod args;
 mod config;
 mod domain_store;
 mod errors;
@@ -46,68 +45,7 @@ mod eviction;
 mod pdns;
 mod routes;
 
-use domain_store::DomainDb;
-
-const USAGE: &'static str = "--host=[host]           'Set local hostname.'
---port=[port]           'Set port to listen on for http connections.'
---cert-directory=[dir]  'Certificate directory.'
---domain=[domain]       'The domain that will be tied to this registration server.'
---dns-ttl=[ttl]         'TTL of the DNS records, in seconds.'
---eviction-delay=[secs] 'How often we purge old records.'
---tunnel-ip=<ip>        'The ip address of the tunnel endpoint'";
-
-const DEFAULT_EVICTION_DELAY: u32 = 120; // In seconds.
-
-struct Args {
-    host: String,
-    port: u16,
-    cert_directory: Option<PathBuf>,
-    domain: String,
-    tunnel_ip: String,
-    dns_ttl: u32,
-    eviction_delay: u32,
-}
-
-impl Args {
-    fn from_matches(matches: ArgMatches) -> Self {
-        let cert_directory = if matches.is_present("cert-directory") {
-            Some(PathBuf::from(matches.value_of("cert-directory").unwrap()))
-        } else {
-            None
-        };
-
-        Args {
-            host: matches.value_of("host").unwrap_or("0.0.0.0").to_owned(),
-            port: value_t!(matches, "port", u16).unwrap_or(4242),
-            cert_directory: cert_directory,
-            domain: matches
-                .value_of("domain")
-                .unwrap_or("knilxof.org")
-                .to_owned(),
-            tunnel_ip: matches
-                .value_of("tunnel-ip")
-                .unwrap_or("0.0.0.0")
-                .to_owned(),
-            dns_ttl: value_t!(matches, "dns-ttl", u32).unwrap_or(60),
-            eviction_delay: value_t!(matches, "eviction-delay", u32)
-                .unwrap_or(DEFAULT_EVICTION_DELAY),
-        }
-    }
-
-    // Gets the args from the default command line.
-    fn new() -> Self {
-        Args::from_matches(App::new("registration_server")
-                               .args_from_usage(USAGE)
-                               .get_matches())
-    }
-
-    // Gets the args from a string array.
-    fn from(params: Vec<&str>) -> Self {
-        Args::from_matches(App::new("registration_server")
-                               .args_from_usage(USAGE)
-                               .get_matches_from(params))
-    }
-}
+use args::Args;
 
 fn main() {
     env_logger::init().unwrap();
@@ -116,13 +54,7 @@ fn main() {
 
     info!("Managing the domain {}", args.domain);
 
-    let config = config::Config {
-        domain_db: DomainDb::new("domains.sqlite"),
-        domain: args.domain,
-        tunnel_ip: args.tunnel_ip,
-        dns_ttl: args.dns_ttl,
-        eviction_delay: args.eviction_delay,
-    };
+    let config = args.to_config();
 
     eviction::evict_old_entries(&config);
 
@@ -159,34 +91,3 @@ fn main() {
 }
 
 // TODO: add iron tests.
-
-#[test]
-fn options_are_good() {
-    let args = Args::from(vec!["registration_server", "--tunnel-ip=1.2.3.4"]);
-
-    assert_eq!(args.port, 4242);
-    assert_eq!(args.host, "0.0.0.0");
-    assert_eq!(args.domain, "knilxof.org");
-    assert_eq!(args.cert_directory, None);
-    assert_eq!(args.tunnel_ip, "1.2.3.4");
-    assert_eq!(args.dns_ttl, 60);
-    assert_eq!(args.eviction_delay, DEFAULT_EVICTION_DELAY);
-
-    let args = Args::from(vec!["registration_server",
-                               "--host=127.0.1.1",
-                               "--port=4343",
-                               "--domain=example.com",
-                               "--cert-directory=/tmp/certs",
-                               "--dns-ttl=120",
-                               "--tunnel-ip=1.2.3.4",
-                               "--eviction-delay=60"]);
-
-    assert_eq!(args.port, 4343);
-    assert_eq!(args.host, "127.0.1.1");
-    assert_eq!(args.domain, "example.com");
-    assert_eq!(args.cert_directory, Some(PathBuf::from("/tmp/certs")));
-    assert_eq!(args.tunnel_ip, "1.2.3.4");
-    assert_eq!(args.dns_ttl, 120);
-    assert_eq!(args.eviction_delay, 60);
-
-}
